requirements.txt
abydos==0.5.0; python_version >= "3.5" and python_version < "4.0"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
black==22.3.0; python_full_version >= "3.6.2"
cachetools==5.1.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
click==8.1.3; python_version >= "3.7" and python_full_version >= "3.6.2" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.4.0" and python_version >= "3.7" and python_version < "4")
colorama==0.4.4; platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.6.2"
deprecated==1.2.13; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
deprecation==2.1.0; python_version >= "3.5" and python_version < "4.0"
et-xmlfile==1.1.0; python_version >= "3.6"
flake8==4.0.1; python_version >= "3.6"
flask-cors==3.0.10
flask-expects-json==1.7.0
flask-jwt-extended==4.4.0; python_version >= "3.6" and python_version < "4"
flask-sock==0.5.2; python_version >= "3.6"
flask-sqlalchemy==2.5.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flask==2.1.2; python_version >= "3.7"
google-api-core==2.7.3; python_version >= "3.6"
google-api-python-client==2.48.0; python_version >= "3.6"
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1; python_version >= "3.6"
google-auth==2.6.6; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
googleapis-common-protos==1.56.1; python_version >= "3.6"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.13.0; python_full_version >= "3.7.0" and python_version >= "3.6"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
itsdangerous==2.1.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.4.0" and python_version >= "3.7" and python_version < "4"
jarowinkler==1.1.2; python_version >= "3.6"
jellyfish==0.9.0; python_version > "3.6"
jinja2==3.1.2; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.4.0" and python_version >= "3.7" and python_version < "4"
jsonschema==4.5.1; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
mccabe==0.6.1; python_version >= "3.6"
mypy-extensions==0.4.3; python_version >= "3.6" and python_full_version >= "3.6.2"
mypy==0.942; python_version >= "3.6"
numpy==1.22.3; python_version >= "3.8"
oauthlib==3.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
openpyxl==3.0.9; python_version >= "3.6"
packaging==21.3; python_version >= "3.6" and python_version < "4.0"
pandas-stubs==1.2.0.58
pandas==1.4.2; python_version >= "3.8"
pathspec==0.9.0; python_full_version >= "3.6.2"
platformdirs==2.5.2; python_version >= "3.7" and python_full_version >= "3.6.2"
protobuf==3.20.1; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pycodestyle==2.8.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pyflakes==2.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pygithub==1.55; python_version >= "3.6"
pyjwt==2.4.0; python_version >= "3.6" and python_version < "4"
pynacl==1.5.0; python_version >= "3.6"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6" and python_version < "4.0"
pyrsistent==0.18.1; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
python-dotenv==0.20.0; python_version >= "3.5"
python-levenshtein==0.12.2; python_version >= "3.5"
pytz==2022.1; python_version >= "3.8"
pyxdameraulevenshtein==1.7.0; python_version >= "3.5"
rapidfuzz==2.1.3; python_version >= "3.6"
requests-oauthlib==1.3.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
shelljob==0.6.3; python_version >= "3.6"
simple-websocket==0.5.2; python_version >= "3.6"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.8"
sqlalchemy==1.4.36; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
textdistance==4.3.0; python_version >= "3.5"
tomli==2.0.1; python_version < "3.11" and python_full_version >= "3.6.2" and python_version >= "3.7"
types-requests==2.27.25
types-urllib3==1.26.14
typing-extensions==4.2.0; python_version >= "3.7"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
werkzeug==2.1.2; python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.4.0" and python_version >= "3.7" and python_version < "4")
wrapt==1.14.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
wsproto==1.1.0; python_full_version >= "3.7.0" and python_version >= "3.6"
